+++ src/main/java/org/cojen/maker/ClassMaker.java
@@ -242,6 +242,15 @@ public interface ClassMaker {
     public ClassMaker addInnerClass(String className);
 
     /**
+     * Add an annotation to this class.
+     *
+     * @param annotationType name or class which refers to an annotation interface
+     * @param visible true if annotation is visible at runtime
+     * @throws IllegalArgumentException if the annotation type is unsupported
+     */
+    public AnnotationMaker addAnnotation(Object annotationType, boolean visible);
+
+    /**
      * Set the source file of this class file by adding a source file attribute.
      *
      * @return this