+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v1.0.3
 ------
 * Default class name when given a lookup must be in the same package.
 * Strict check when adding constructor with a MethodType.
+* Allow methods defined in Object to be invoked on interface instances.
 
 v1.0.2 (2021-05-05)
 ------