+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 Changelog
 =========
 
-v1.0.3
+v1.1.0
 ------
 * Default class name when given a lookup must be in the same package.
 * Strict check when adding constructor with a MethodType.
 * Allow methods defined in Object to be invoked on interface instances.
+* Support adding annotations.
 
 v1.0.2 (2021-05-05)
 ------